Clear stdin before reading user input
I often repeatedly press enter in my terminal while scap is in progress,
or I might mistakengly paste the clipboard in the terminal. The
erroneous input ends up being read and that eventually leads to abort an
ongoing deployment.

When running interactively, read the whole of stdin before invoking
input().