Clear stdin before reading user input (take 2)
I often repeatedly press enter in my terminal while scap is in progress,
or I might mistakengly paste the clipboard in the terminal. The
erroneous input ends up being read and that eventually leads to abort an
ongoing deployment.

This was previously made in 25227595 which used `sys.stdin.read()` to
consume the pending input, that blocks until a newline is available
which is only the case when something has been sent.