Drop -1 from version string
The `-1` at the end of the scap version string seems to be a holdover
from when scap was Debian-packaged.  We no longer use Debian packaging
for scap so this extra bit of version is unnecessary.  It also
generates a warning when building wheels:

```
/tmp/build-env-z54ckp5t/lib/python3.11/site-packages/setuptools/dist.py:294: InformationOnly: Normalizing '4.110.0-1' to '4.110.0.post1'
  self.metadata.version = self._normalize_version(self.metadata.version)
```

This commit drops the `-1` from the current version string and avoids
re-adding it in the release scripts.

Change-Id: Iff2f5616b6e4f4dc305588e697f11964bd2df0a0