tests/scap/test_cli.py: Fix test_announce
Some future code that I will be adding to `announce()` will generate a
warning if the number of supplied parameters does not match the number
expected by the format string.  I changed a test call to `announce()`
which was passing a parameter that was not accessed by the format
string.

Before:
```
app.announce("test", "bar")
```

After:
```
app.announce("test %s", "bar")
```

Change-Id: I1788e2d2392674d9d42423203f2c83101a11959e