web: Add support for VITE_APISERVER_URL
If `VITE_APISERVER_URL` is set in `web/.env.development`, it will be
used as the base URL for apiserver accesses when not in test mode.  If
`VITE_APISERVER_URL` is unset (the default), the window origin is used
as the base URL.

In the process, changed `makeApiUrl()` into a method on the Pinia
store added the `apiserverBaseURL` getter.

Change-Id: I501311c59f6182c4ae9c9044f4b18972d586366d