Spiderpig: Add output_line interaction with sensitive flag
Initial support for `output_line()` interaction, which has a flag to
indicate if the information is sensitive.  For `TerminalIO` the line
is simply printed.  For `SpiderpigIO` the line and sensitive flag are
passed through to the jobrunner which currently just logs the
sensitive lines in a prominent way.  A followup commit will change the
jobrunner to log in a way that allows the sensitive information to be
excluded during log retrieval.

Use `output_line()` in `scap/patches.py` to output the sensitive
information about what patches are being applied.

Change-Id: Ie937da9e88863a9e699df2d87c272e8a8f01fdb6