web: Get rid of remaining eslint warning
.. by defining the `Interaction` interface (translated from
`scap/spiderpig/model.py`) and using it as the type of the
`interaction` prop of the `SpInteraction` component.  Thanks to Eric
Gardner for assistance!

Since this added the first TypeScript file to the project, I had to
tweak `.eslintrc.json` to add TypeScript support and update the
package list in `package.json` accordingly.

Add `npm run test` to the Gitlab CI pipeline

Change-Id: Ic9101286a917f13e7c1d66721bd9b41c1ebfabfe