Reduce the cardinality of scap timers/statsd metrics (phase 1)
This commit plucks the low hanging fruit of the `scap deploy` and
`scap prep` timers.

log.py:

Allow the timer measurement name to be specified separately from the
description that is displayed to the user.  Also allow transmission of
the measurement to be disabled by specifying a name of `None`.

cli.py:

Update the signature of `Application.Timer()` to match the changes
made to `log.Timer()`.

deploy.py:

Specify a measurement name of `deploy`.

prep.py:

Specify a measurement name of `prep`.

Bug: T377883
Change-Id: I56b7ebbc58ed1460f658a2049586d57d69944a52