web: add a JobCard
Converts the Job History list from a table to cards with job data.
JobCard uses CdxCard to display the job information.

- Replace the CdxTable of job data with a list of JobCard components.
- Remove "Queued" and "ChangeId" labels from job history overview page.
- Use `v-bind` to pass the Job object properties as fallback attributes
then define each job object property as props.
- Add InfoChips to display the status message.
- Job id links navigate to the job details page.
- We decided the cards will not be clickable for now because any click
inside the card caused navigation to the JobDetails page. When you select
an interaction choice displayed inside the JobCard, the click responded
to the interaction prompt and navigated to the JobDetails page.
- Highlight a job card that needs attention.
- Display the job status in the card for a job in progress.
- Interaction prompt and choices are displayed on main page in the JobCard.
The interaction prompt content overflows so I applied a vertical scroll.
- Add column labels that look similar to table headers on larger screens.

Bug: T379412