scap clean: first delete branch from superproject
mediawiki/core wmf branch is a super project of extensions, skins,
vendor. When deleting the branch from one of the subproject repository,
Gerrit fails to format the commit object that it wants to craft for the
parent project. That is due to the branch object being deleted which is
not supported by Gerrit 3.10.2.

Delete the branch from the superproject before processing with the
tracked repositories.

Bug: T303828
X-Easter-Egg: inline comment