Remove check for agreement between all.dblist and wikiversions.json
This conflicts with what I'm doing with preinstall.dblist and makes the
currently merged deployment branches undeployable.

The right place for this sort of check is the mediawiki-config PHPUnit
test DbListTest. I already fixed testDblistAllContainsEverything there.

Back story:

For a couple of months I have been working on a rewrite of addWiki.php.
The main thrust of the change is to configure the wiki first, and then
use that configuration to install the database and tables, instead of
using a lot of special case logic in WikimediaMaintenance. (T352113)

Consequently, we need to be able to configure wikis in mediawiki-config
when they don't have a database yet. There are a number of ways that
could be done, but I decided that the most backwards compatible way
would be for all.dblist to continue to reflect the installed and usable
wikis, while a new preinstall.dblist would contain wikis that are
configured but not yet usable.

The main mediawiki-config component of this change was done on November
28 in I47adcb776cf411244a4aa1591d1333acddd90f69.

Now that the installer project is nearing completion, I would like to do
a production test of the new system by creating id.wikivoyage.org. The
first step towards that is deploying a change which adds idwikivoyage
to wikiversions.json but not to all.dblist.
(T380726, Ia074b84f9d3d4425f1416e98928b74135e4806bd)

scap fails to deploy this change, complaining that not all wikis are in
all.dblist, but that is intentional. I already fixed a mediawiki-config
unit test complaining about the same thing in
If8e93120a3cb70eec83ac692ffa3089d2f881813.

Bug: T352113