web: improve JobViewerPage UI
On the JobViewerPage UI, the user can interact with the
interaction prompt by choosing a button action. The job
card displays additional info like queued time and terminal
log. The "Close Log" button navigates to the main page.

- Migrate the terminal log, sensitive data checkbox, and button
actions to a new component. The buttons "Interupt Job" and "Kill
Job" are now nested in a CdxMenuButton. The former button ids,
"kill" and "interrupt" match the value defined in MenuButton's
menu item.
- Remove job summary content because it repeated information
already displayed in the job card.

Bug: T379413

Closes T379413