JobCard: Make the cards into clickable links
The Codex Card component (which is used inside JobCard here)
can behave like a clickable link if a "url" prop is provided.

This change enables this behavior by using vue-router's
router.resolve() function, which returns a normalized route
location (including a simple href string) for a set of
parameters.

To ensure that we only make cards linkable when the user is
not already on the page corresponding to the job in question,
this logic is housed inside a new "jobLink" computed property
in the JobCard component. If the resolved link differs from the
current route link, then it is returned (and passed to the CdxCard
component in the template); otherwise null is returned.

Bug: T379413