main.py: Slight improvement to retry/continue/exit prompt
`AbstractSync.retry_continue_exit()`:

Include the fact that the operation failed in the prompt.  When
running scap in a terminal, the error output will be right above the
prompt, so it's obviously already, but when running under SpiderPig,
the job log is not displayed by default, so it's useful to have a
summary of the problem in the prompt.

Change-Id: I920ebad63b46e2c7ca17bbd3de97becbefe28d76