kubernetes: support non-deploy helmfile releases
A non-deploy release is one for which scap only manages release value
files and does not deploy to (or otherwise interact with) k8s.

This is convenient for use cases such as mw-script, where releases are
meant to follow the most recently deployed images [0], but are deployed
externally (in this case, manually).

More broadly, any use case for which deployment is asynchronous in this
way could benefit from non-deploy releases.

A release can by marked non-deploy by setting `deploy: false` in
mediawiki-deployments.yaml on the release configuration definition, as
described in the now-extended DeploymentsConfig docstring.

[0] This currently happens by way of reading the release values for
mw-web's main release, which is non-ideal.

Bug: T387917