Makefile: change test image namespace to localhost
When building the image through podman and with docker.io/moby/buildkit,
the image is named `docker.io/local/scap-bullseye-test` and running the
test referring to it as `local/scap-bullseye-test` fails since the image
can not be found.

Changing the namespace to `localhost` prevents the automatic prefixing
of `docker.io`.