Makefile: explicitly output the image
When building the image through Podman and with docker.io/moby/buildkit,
it is not exported. Running the test referring to the image thus fails
since the image is not available.  I think `docker buildx` does the
export to the Docker daemon, but with Podman that has to be explicitly
requested.