spiderpig apiserver: Change log job transmission format
Instead of returning a job log as a stream of characters, return it as
a stream of JSON lines, one record per log line.  Also include some
meta records which identify operational boundaries.  In particular,
add an EOF record that lets the client know that it has definitely
reached the end of a job log and that there is no further data to
read.  This allows the client to detect when the job log stream has
been prematurely terminated.

Change-Id: Ic3ab9ab16e0b0b1e888b995ca014f2f648c69500