feat: Add Logstash link in the nav bar
Today, during an early SpiderPig test drive, the Release Engineering
team talked about adding an ambient error rate indicator to SpiderPig.

As a first step, a link to Logstash may be a good start. In future, this
may be replaced by some indication of the current error rate with a link
to the appropriate dashboard.

Change-Id: Id61e0c1d3f77d741f98a596864d6c50aab20f28c