utils.get_real_username: Use SPIDERPIG_REAL_USER if available
`utils.get_real_username()` is now defined to return the (shell) name
of the user that initiated the current operation.  If
`SPIDERPIG_REAL_USER` is set (which happens when spiderpig-jobrunner
starts a job), its value will be returned as the real user.

Analysis of places where `utils.get_real_username()` is used:

* `git.set_env_vars_for_user()`, to set the author of a commit message.
* In `lock.py` to record the lock holder/releaser.
* In `log.py` to emit `!log user@host` IRC messages and in logstash records.
* In `backport.py`, used when approving a Gerrit change or creating a
  revert commit.
* In `main.py`, `AbstractSync`: `_init_history()` and
  `_pause_after_testserver_sync`.  These previously used
  `getpass.getuser()`.

Change-Id: I9d719cad3b9fb3b8e3a28d064dd297033a1edad7