kubernetes: support image "kind" and deprecate the debug boolean
This patch does two, closely related things:

First, it adds the mw_kind field to DeploymentsConfig items.

At a high level, "kind" is analogous to an image build target - i.e., a
procedure / recipe that is used by the make-container-image scripts to
produce an image. Think, e.g., the debug image.

In contrast, "flavour" continues to represent the particular combination
of parameters used to instantiate the image - e.g., build args, like the
PHP version).

To minimize coupling between scap and make-container-image, mw_kind is
treated as an opaque key into the build report (e.g., "debug-image"),
very similar to how flavour is already treated.

To compensate for the fact that we're introducing "another opaque key to
get wrong", we now collect all helmfile release values updates before we
attempt to apply any of them, so a misconfiguration is surfaced early.

This happens *after* image build in order to avoid surprises in the case
where mediawiki-deployments.yaml is updated at the time of the first
scap run after a new kind is added.

Second, it begins deprecation of the debug boolean, in favor of
explicitly selecting the debug-image kind and testservers stage.

Once mediawiki-deployments.yaml and train-dev's equivalant are updated,
I'll remove support for debug.

Bug: T389499