spiderpig: auto select first backport search match
Replace the Codex cdx-multiselect-lookup component with a Vuetify
v-combobox which supports an `auto-select-first` property that
highlights the first option and selects it on blur. This saves folks
from needing to hit the down key to highlight the option. We also
configure `,` and ` ` as delimiters in the input field to make data
entry match user intuition seen during tests with the RelEng team.

Change-Id: Id31ce1e892a59419ad5e66da67d3aaa0093b97ad