Allow restarting multiple services when require_valid_service is True
When declaring multiple services in service_name and if
require_valid_service is True the first missing/masked service will
cause the process to stop attempting to restart subsequent services.
For deployments with multiple services to restart and where the cluster
is heterogeneous we would like scap to continue restarting services
event if a single one is missing.

Relates to T221709 where we want to restart mulitple services but the
service names might vary depending on the machine role:
- wdqs-blazegraph vs wcqs-blazegraph
- wdqs-categories not always setup