spiderpig: Add means to prepopulate backport change numbers
If one or more `backport=<some change number>` query parameters are
added to the URL of the overview page, then the corresponding change
numbers will be prepopulated in the change numbers box of the Backport
component.

Also added an error dialog if starting a backport fails due to bad
input.

Also:
* Renamed a couple of variables to better describe what they hold.

Change-Id: Ie0e4f4649ca6b80c20c38cae2a7053becbf2c823