spiderpig: Fix auth loop bug
On the SpiderPig Overview page there are two timers that make API
calls every second.  When `api.js` encounters an authentication error
while making a request, it makes the browser navigate to the page
supplied by the server (this will be the IDP server URL).  It's
possible that while processing the navigation for one timer's auth
failure, auth failure for the other timer could happen, resulting in
the prior navigation being cancelled and a new one started.  If the
navigation doesn't complete within one second, this renavigation can
continue on forever.

Fix this by adding a `setWindowLocation()` function which establishes
as "first writer wins" strategy for navigations.  Changed all direct
sets of `window.location.href` to use `setWindowLocation()`.

Change-Id: I756712b00dec199fcd3ef154b83d35f2fd5ddb82