log.py: @version should be "1"
The `@version` field of generated logstash records should be the
string `"1"`, not an integer.  This matches all examples seen in
https://docs.opensearch.org/docs/latest/tools/logstash/index/.  I
noticed that in OpenSearch Dashboards, the `@version` field was
indicating `Conflicting field`, so I googled on the subject and came
to the conclusion that the mismatched type was the issue.

Change-Id: I4118c765fd379e8585ffe58ebfb3764fc19f8f6e