web/src/state.js: Ensure that this.backportJob is always a string
https://developer.mozilla.org/en-US/docs/Web/API/Window/localStorage#description
says that LocalStorage keys and values should be strings, so make sure
we always convert `this.backportJob` to a string when writing it and
convert back to an integer (or null) when reading it.  For consistency
with the rest of the code in the file, use
`JSON.stringify`/`JSON.parse` to achieve this.

Bug: T394033
Change-Id: I37384affbef45d9347a98e0ef01624c108b57a6f