spiderpig: Handle simple add/drop of table columns
Allow for automatic handling of added or dropped table columns.  If a
model defines a column that does not exist in a table, the column is
added automatically.  Likewise, if a table has a column that is not
defined in its corresponding model, drop the column.

If a `NOT NULL` column is added, a `server_default` value must be
configured.

Schema changes are processed when spiderpig-jobrunner starts up.

Note that this does not handle things such as the configuration of a
column changing.

Change-Id: I055cc81307cea17c5808a5ea00e307b5d823061a