Update scap beta release scripts
Host names are out of date.

The `--batch` flag for `scap install-world` is now the `--yes` flag and
install-world fails if it's omitted.

The `install_local_version` script now requires a tag, use git to find
the last five tags and prompt the user.

Interaction now looks like this:

```
$ ./release-scripts/update-scap-in-beta
No scap verison specified!

Pick a scap version to deploy:
     1  4.99.0
     2  4.98.0
     3  4.97.0
     4  4.96.0
     5  4.95.0
Enter the number of the version above (empty or 0 to cancel) → hi
    'hi' is invalid, please enter a number between 0 and 5
Enter the number of the version above (empty or 0 to cancel) → 6
    '6' is invalid, please enter a number between 0 and 5
Enter the number of the version above (empty or 0 to cancel) → 4

    Selected version '4.96.0'

Use this log message in #wikimedia-releng IRC:

    !log Upgrading scap to 4.96.0 in beta cluster

Press any key to deploy (Ctrl-C to cancel)
^C
```

Co-authored-by: Jaime Nuche <jnuche@wikimedia.org>
Change-Id: I193fa960a051f39243e5d46d29d88034e7d7c589