Drop several php_fpm* config parameters
The following scap config parameters are no longer used (and will be
ignored if supplied):

* php_fpm_opcache_threshold
* php_fpm_always_restart
* php_fpm

Previously, the values of these config settings were passed as
arguments to `php_fpm_restart_script`, but the restart script used in
beta and production (and now train-dev) does not use the arguments, so
they have been dropped and the corresponding code simplified.

Also, `php_fpm_unsafe_restart_script` is deprecated and will soon be removed.

Bug: T396166
Change-Id: Iaed455c047adf9de2271de7327e935057c7ad41d