spiderpig/api.py: Include Content-Type header in job log response (v2)
Include a Content-Type header in the `/api/jobs/{job_id}/log` response
to make it work better with Firefox.  Without a Content-Type header,
Firefox will wait for a few hundreds of bytes of data to arrive before
delivering it to Javascript code.

The Content-Type _should_ be `application/jsonl` but we use
`application/octet-stream` since that's the only content-type that the
proxy servers will accept without doing weird things to the
transmission (such as collecting all the chunks before transmitting to
the browser).

Bug: T397097
Change-Id: I3f9f8030532957597728f9830c06615223294106