kubernetes: adopt new report structure and set php.version
This takes advantage of the change in make-container-image to imbue
flavours with metadata describing the resulting images, and begins using
the first metadata attribute: the PHP version.

The latter is used to populate the php.version helmfile value, which is
a significant improvement over having to coordinate updates to the value
in parallel with changes to the release-to-flavour mapping.

See release!207 for additional benefits of this approach over using
image metadata labels to expose this information, the shortcomings of
which led 02b817a7a375a6df7b2edfaec925cd7e4b610ea7 to be reverted in
b0b463503df27a1ba5229ae3fd6933dceae4798a.

Bug: T401721