spiderpig: Avoid race condition when refreshing train status
The train status is retrieved from multiple sources, the
`/api/train/status` call which reflects the status according to the
deployment server's on-disk state, and the memoized (original) status
stored in any train job currently being processed by the jobrunner. The
latter is needed to accurately present the effects of an in-progress
train job.

Consulting the two sources completely independently (one in `refresh()`
and the other in `watch( jobrunner.status, ... )`) results in a race
condition. If the former fires last and overwrites the status of the
prior, the in-progress train status is inaccurately presented.

To avoid the race (or at least its effects), have `refresh()` also fetch
the jobrunner status and prefer the memoized train status value. This is
slightly inefficient network wise since two jobrunner status requests
are in flight but the inefficiency is only incurred during the component
mount.

Bug: T399120
Change-Id: I32b7e3c2c12c116a8ab8c04c89a6134010763a0a