utils.py: Make select_latest_patches treat /srv/patches/next as latest
Changed `select_latest_patches` to filter candidate directories using
`parse_wmf_version` (which handles versions such as "next").
Previously it would only consider `x.yy.z-wmf.w` style versions.

This means that if `/srv/patches/next` exists, it will be used by
`scap prep` as the source to populate `/srv/patches/<version>`.  This
is the naturally correct thing to do for T398873 and is also the
correct behavior according to the discussion in T295925.

Bug: T398873
Bug: T295925
Change-Id: I43ed1268c44e81a65c7333472f08102c17da48cf