spiderpig: Better support days leading up to train
Display an "arriving" status in the calendar card at the beginning of a
given train week when the version of the task doesn't yet match the
current train version. The latter is derived from the branch in Gerrit
and the branch isn't created until the date of the task (currently
Tuesday of each week).

Once 0300 UTC on the date of the task has passed, display a warning if
the versions still do not match.

Change-Id: I24c6ec5c9fcdd8bd0a2aba9a1dfc21c23f13ed74