Teach apply-patches about submodules
If an extension or skin has submodules, then we should be able to add
patches to the submodule by adding them to /srv/patches.

We can add the patch files such that the relative path to the patch is
the same as the relative path to the submodule in the extension.

Example: the `Body` exension with the `Foot` submodule:

    Body
    ├── Lower
        ├── ...
    │   └── Foot  <- (submodule)
    └── Upper
        └── ...

On the deploy server, we want to patch `Toe.php` in the Foot submodule:

    /srv/mw/php-<v>/extensions/Body/Lower/Foot/Toe.php

To do that, we'd add a patch:

    /srv/patches/<v>/Extensions/Body/Lower/Foot/01-T1000.patch

Change-Id: Ica8b03600efe46b97414206a49ed55a54cd8e2cc