docs: Drop sphinxcontrib.{actdiag,blockdiag}, use sphinx.ext.graphviz
`sphinxcontrib.actdiag` and `sphinxcontrib.blockdiag` have not been
updated since 2021 and both have a dependency on a version of `Pillow`
that has vulnerabilities.  So drop the use of `sphinxcontrib.actdiag`
and `sphinxcontrib.blockdiag`, and use `sphinx.ext.graphviz` instead.

`.pipeline/blubber.yaml`: Add the `graphviz` package which is now
needed to build docs.

Change-Id: I9c94f2fd1b0036a04b031ab640790181e31f6556