kubernetes.py: prune_local_images: don't remove parent images
Don't delete parent images of live incremental images.  This avoids
triggering full image builds when `scap clean-images` is followed by
`scap sync-world`.

This relies on a label added by https://gitlab.wikimedia.org/repos/releng/release/-/merge_requests/202

I removed the parallel execution code for simplicity.  It would be
reasonable to revive it for the image removal phase of the operation,
but parallel operation is not necessary for the mark phase.

Change-Id: I85558cb9d5e5ef948b6154f93953b03f780f76c8