git.py: Fix info() public commit selection
Previously `git.info()` used `git merge-base HEAD origin` to find the
public commit for a given repo.  The problem with this approach is
that `origin` resolves to the commit at the tip of the default branch
of the remote repo.  But we don't want to be comparing with the
remote's default branch.  We want to compare with the branch that
corresponds to the branch that the local checkout is using.

Deal with this by allowing the caller to supply the name of the branch
that should be compared.

Bug: T405738
Change-Id: I378b7ff36ab2d84f145818e8e29265db2d641278