mwscript.py: Log stdout on CalledProcessError
In `run` and `run_shell`, if the subprocess fails with
`CalledProcessError` log any non-blank stdout in addition to stderr.

It may turn out that showing stdout may be inappropriate for some
error scenarios.  If that is the case we can follow up with a keyword
argument that allows control over whether stdout/stderr are logged
when `CalledProcessError` is raised.

Bug: T396169
Change-Id: I912a383f9aa0c88ff943a43e9fe25d8ca94aa923