main.py: Improve wording of retry_continue_exit exit case
Update the wording and shortcut keys used when prompting to
retry/continue/exit. The continue case is now labeled "Ignore failure
and proceed with deployment" which hopefully makes the potential impact
of that branch clearer to the deployer. The exit case is now labeled
"Cancel deployment" again in the hope that this wording makes the impact
clear to the deployer.

These changes will hopefully make a bit more sense for deployers who
have only had experience in using the SpiderPig web GUI.

Change-Id: I85c804ecb8f50254f50aa3d7ac4c016276b336d9